feat: 유저 정보 추가 #32

+++ src/main/java/com/gdg/z_meet/domain/meeting/dto/MeetingResponseDTO.java
@@ -19,7 +19,7 @@ public class MeetingResponseDTO {
         String name;
         Integer verification;
         List<String> major;
-        Integer age;
+        Double age;
         List<String> music;
     }
