+++ src/main/java/com/gdg/z_meet/domain/meeting/repository/UserTeamRepository.java
@@ -10,5 +10,5 @@ public interface UserTeamRepository extends JpaRepository<UserTeam, Long> {
 
     boolean existsByUserIdAndTeamId(Long userId, Long teamId);
     List<UserTeam> findByTeamId(Long teamId);
-    Integer countByTeamId(Long id);
+    Long countByTeamId(Long id);
 }
